Build wayland support optionally
